libmach: Use long for offset
As we use standard fseek() and ftell() we cannot support
files that cannot be addressed with longs, so it is better
to check for these problems when reading files and use
long after that point.

+++ include/bits/scc/mach.h
@@ -84,7 +84,7 @@ struct section {
 	unsigned long long base;
 	unsigned long long load;
 	unsigned long long size;
-	unsigned long long offset;
+	long offset;
 
 	unsigned flags;
 	int index;
@@ -111,7 +111,6 @@ struct map {
 	struct mapsec *sec;
 	struct mapsec *seg;
 };
-
 
 /**
  * @stype: Used internally by libmach
